src/notemap-mcp/requirements.txt
mcp>=1.20,<2.0
python-frontmatter>=1.1.0
python-slugify>=8.0.0
pyyaml>=6.0
